==> scripts/3.1-ClimateFireSummerAnalysis.Rmd <==
---
title: "Climate Fire Summer"
author: "Anna Talucci"
date: "6/20/2021"
output: html_document
---




# Clear workspace

```{r}
rm(list=ls()) 
```


# Overview
[Terra Climate](https://developers.google.com/earth-engine/datasets/catalog/IDAHO_EPSCOR_TERRACLIMATE#bands)

Variables 
DEF
Precip

# Library
```{r include=FALSE}

library(dplyr)
library(tidyr)
library(ggplot2)
library(cowplot)
library(tidyverse)
library(broom)
library(ggplot2)
library(stringr)
library(lme4)
library(cowplot)
library(lme4)
library(lsmeans)
library(magrittr)
library(rcompanion)
library(qpcR)
library(tibble)
```

# Define graph Theme and color Palette 

##Themes

```{r}

reg_theme = theme_bw() + theme(legend.position = "none") +
  theme(plot.margin = unit(c(t = 0.3, r = 0.3, b = 0.3, l = 0.1), "cm")) +
  theme(axis.title.x = element_text(size = 10, hjust = 0.5, vjust = -0.2),
        axis.title.y = element_text(size = 10, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 9, color = "black"),
        axis.text.y = element_text(size = 9, color = "black"),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "black"))

```

## Palettes


```{r}
cbbBTCodesPalette <- c("#009E73", "#0072B2", "#000000", "#E69F00", "#56B4E9", "#F0E442", "#CC79A7", "#D55E00")
```

4. EST = B1 = "#009E73",
6. NST = B2 = "#0072B2"
1. BT = T1 = "#000000"
2. CKMT = T2 = "#E69F00"
3. CPT = T3 = "#56B4E9"
5. NSCT = T4 = "#F0E442"
8. TCST = T5 = "#CC79A7"
7. TBBMT = T6 = "#D55E00"



# Data 

## Climate
```{r}
climate = read.csv("../data/ClimateFire/ClimateFire2001-2020.csv") 
```

```{r}
climate1 = climate %>% 
  dplyr::select(EcoCode:MHa, SummerPr, SummerDef, SummerTmmx) %>%
  mutate(log10SizeHa = log10(SizeHa+.1)) %>%
  mutate(BTCodes = ifelse(EcoCode =="EST", "B1",
                  ifelse(EcoCode =="NST", "B2",
                   ifelse(EcoCode =="BT", "T1",
                       ifelse(EcoCode =="CPT", "T2",
                         ifelse(EcoCode =="CKMT", "T3",
                            ifelse(EcoCode =="NSCT", "T4",
                              ifelse(EcoCode =="TCST", "T5", "T6"))))))))

climate1
```

## Snow
```{r}
snow = read.csv("../data/SnowData/SnowOff2001-2020.csv") 
```

```{r}
snow1 = snow %>% 
  mutate(log10SizeHa = log10(SizeHa+.1)) %>%
  mutate(BTCodes = ifelse(EcoCode =="EST", "B1",
                  ifelse(EcoCode =="NST", "B2",
                   ifelse(EcoCode =="BT", "T1",
                       ifelse(EcoCode =="CPT", "T2",
                         ifelse(EcoCode =="CKMT", "T3",
                            ifelse(EcoCode =="NSCT", "T4",
                              ifelse(EcoCode =="TCST", "T5", "T6"))))))))

snow1

```

## Fire 

```{r}
fire = read.csv("../data/1.0-DataPrep/SiberiaFires2001-2020.csv") 
```

```{r}
fire 
```
```{r}
fire1 = fire %>%
  group_by(FireYr, EcoName, EcoCode) %>%
  summarise(start = round(min(MinDay), 0),
            end = round(max(MaxDay), 0),
            SizeHa = (sum(SizeHa))) %>%
  mutate(season = end-start) %>% 
  mutate(log10SizeHa = log10(SizeHa+.1)) %>%
  mutate(BTCodes = ifelse(EcoCode =="EST", "B1",
                  ifelse(EcoCode =="NST", "B2",
                   ifelse(EcoCode =="BT", "T1",
                       ifelse(EcoCode =="CPT", "T2",
                         ifelse(EcoCode =="CKMT", "T3",
                            ifelse(EcoCode =="NSCT", "T4",
                              ifelse(EcoCode =="TCST", "T5", "T6"))))))))

fire1

```

# Analysis  Summer Variables
## Water Deficit Model
```{r}
fit1 = lm(log10SizeHa ~ SummerDef + EcoCode + SummerDef:EcoCode, data = climate1)
```

### Residuals


```{r}
# Get the fitted value for each observation
climate1$fit1 = fitted(fit1)
```

```{r}
# Get the residuals of the model
climate1$res1 = residuals(fit1)
```

```{r}
qplot(x = fit1, y = res1, data = climate1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = SummerDef, y = res1, data = climate1,
xlab = "CMD",
ylab = "Residuals",
main = "Residuals vs CMD")
```
```{r}
qplot(x = factor(1), y = res1, data = climate1, geom = "boxplot")
```

```{r}
qqnorm(climate1$res1, main = "Normal QQ Plot of Residuals")
qqline(climate1$res1) # add reference line to the qq plot
```

```{r}
plot(fit1, which = 1) # residual vs fitted values
```

```{r}
plot(fit1, which = 2) # qqnorm plot of residuals
```

### Summary
```{r}
summary(fit1)
```
```{r}
summary(fit1)$r.squared 
```
```{r}
int1_BT = coef(fit1)[1]
int1_CKMT = coef(fit1)[1] + coef(fit1)[3]
int1_CPT = coef(fit1)[1] + coef(fit1)[4]
int1_EST = coef(fit1)[1] + coef(fit1)[5]
int1_NSCT = coef(fit1)[1] + coef(fit1)[6]
int1_NST = coef(fit1)[1] + coef(fit1)[7]
int1_TBBMT = coef(fit1)[1] + coef(fit1)[8]
int1_TCST = coef(fit1)[1] + coef(fit1)[9]


slope1_BT = coef(fit1)[2]
slope1_CKMT = coef(fit1)[2] + coef(fit1)[10]
slope1_CPT = coef(fit1)[2] + coef(fit1)[11]
slope1_EST = coef(fit1)[2] + coef(fit1)[12]
slope1_NSCT = coef(fit1)[2] + coef(fit1)[13]
slope1_NST = coef(fit1)[2] + coef(fit1)[14]
slope1_TBBMT = coef(fit1)[2] + coef(fit1)[15]
slope1_TCST = coef(fit1)[2] + coef(fit1)[16]

```

### Create Dataframe with summry output
```{r}
EcoCode = c("BT", "CKMT", "CPT", "EST", "NSCT", "NST", "TBBMT", "TCST")
intercept = c(int1_BT, int1_CKMT, int1_CPT, int1_EST, int1_NSCT, int1_NST, int1_TBBMT, int1_TCST)
slope = c(slope1_BT, slope1_CKMT, slope1_CPT, slope1_EST, slope1_NSCT, slope1_NST, slope1_TBBMT, slope1_TCST)

sum_Def =data.frame(EcoCode,intercept, slope)
sum_Def
```

### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
climate2 =climate1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ SummerDef, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results)%>%
  mutate(variable="def") %>%
  left_join(sum_Def) %>%
  dplyr::select(variable, EcoCode, intercept, slope, r.squared:nobs)

climate2
```

```{r eval=FALSE, include=FALSE}
climate2 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ SummerDef, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(EcoCode), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "Climate water deficit", y = expression(R^{2}))
```



### Combo Plot
```{r fig.height=2.3, fig.width=3}
g1 = ggplot(data = climate1, aes(x = SummerDef, y = log10SizeHa, color = BTCodes)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbBTCodesPalette) +
    labs(y=("Annual area burned\n(ha)"), x= "Summer water deficit (mm)", color = "EcoCode") +
  reg_theme

g1
```




## Precipitation Model

```{r}
fit2 = lm(log10SizeHa ~ SummerPr + EcoCode + SummerPr:EcoCode, data = climate1)
```

### Residuals

```{r}
# Get the fitted value for each observation
climate1$fit2 = fitted(fit2)
```

```{r}
# Get the residuals of the model
climate1$res2 = residuals(fit2)
```

```{r}
qplot(x = fit2, y = res2, data = climate1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = SummerPr, y = res2, data = climate1,
xlab = "CMD",
ylab = "Residuals",
main = "Residuals vs CMD")
```
```{r}
qplot(x = factor(1), y = res1, data = climate1, geom = "boxplot")
```

```{r}
qqnorm(climate1$res2, main = "Normal QQ Plot of Residuals")
qqline(climate1$res2) # add reference line to the qq plot
```

```{r}
plot(fit2, which = 1) # residual vs fitted values
```

```{r}
plot(fit2, which = 2) # qqnorm plot of residuals
```

### Summary
```{r}
summary(fit2)
```
```{r}
summary(fit2)$r.squared 
```
```{r}
int2_BT = coef(fit2)[1]
int2_CKMT = coef(fit2)[1] + coef(fit2)[3]
int2_CPT = coef(fit2)[1] + coef(fit2)[4]
int2_EST = coef(fit2)[1] + coef(fit2)[5]
int2_NSCT = coef(fit2)[1] + coef(fit2)[6]
int2_NST = coef(fit2)[1] + coef(fit2)[7]
int2_TBBMT = coef(fit2)[1] + coef(fit2)[8]
int2_TCST = coef(fit2)[1] + coef(fit2)[9]


slope2_BT = coef(fit2)[2]
slope2_CKMT = coef(fit2)[2] + coef(fit2)[10]
slope2_CPT = coef(fit2)[2] + coef(fit2)[11]
slope2_EST = coef(fit2)[2] + coef(fit2)[12]
slope2_NSCT = coef(fit2)[2] + coef(fit2)[13]
slope2_NST = coef(fit2)[2] + coef(fit2)[14]
slope2_TBBMT = coef(fit2)[2] + coef(fit2)[15]
slope2_TCST = coef(fit2)[2] + coef(fit2)[16]



```

### Create Dataframe with summry output
```{r}
EcoCode = c("BT", "CKMT", "CPT", "EST", "NSCT", "NST", "TBBMT", "TCST")
intercept = c(int2_BT, int2_CKMT, int2_CPT, int2_EST, int2_NSCT, int2_NST, int2_TBBMT, int2_TCST)
slope = c(slope2_BT, slope2_CKMT, slope2_CPT, slope2_EST, slope2_NSCT, slope2_NST, slope2_TBBMT, slope2_TCST)

sum_Pr =data.frame(EcoCode,intercept, slope)
sum_Pr
```

### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
climate3 =  climate1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ SummerPr, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results)  %>%
  mutate(variable="pr") %>%
  left_join(sum_Pr) %>%
  dplyr::select(variable, EcoCode, intercept, slope, r.squared:nobs)

climate3
```

```{r}
climate1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ SummerPr, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(EcoCode), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "Precipitation", y = expression(R^{2}))
```


### Combo Plot
```{r fig.height=2.3, fig.width=3}
g2 = ggplot(data = climate1, aes(x = SummerPr, y = log10SizeHa, color = BTCodes)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbBTCodesPalette) +
  labs(y=("Annual area burned\n(ha)"), x="Summer precipitation (mm)") +
  reg_theme

g2
```

## Max Temperature model


```{r}
fit3 = lm(log10SizeHa ~ SummerTmmx + EcoCode + SummerTmmx:EcoCode, data = climate1)
```

### Residuals
```{r}
# Get the fitted value for each observation
climate1$fit3 = fitted(fit3)
```

```{r}
# Get the residuals of the model
climate1$res2 = residuals(fit3)
```

```{r}
qplot(x = fit3, y = res2, data = climate1,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = SummerTmmx, y = res2, data = climate1,
xlab = "CMD",
ylab = "Residuals",
main = "Residuals vs CMD")
```
```{r}
qplot(x = factor(1), y = res1, data = climate1, geom = "boxplot")
```

```{r}
qqnorm(climate1$res2, main = "Normal QQ Plot of Residuals")
qqline(climate1$res2) # add reference line to the qq plot
```

```{r}
plot(fit3, which = 1) # residual vs fitted values
```

```{r}
plot(fit3, which = 2) # qqnorm plot of residuals
```

### Summary
```{r}
summary(fit3)
```
```{r}
summary(fit3)$r.squared 
```
```{r}
int3_BT = coef(fit3)[1]
int3_CKMT = coef(fit3)[1] + coef(fit3)[3]
int3_CPT = coef(fit3)[1] + coef(fit3)[4]
int3_EST = coef(fit3)[1] + coef(fit3)[5]
int3_NSCT = coef(fit3)[1] + coef(fit3)[6]
int3_NST = coef(fit3)[1] + coef(fit3)[7]
int3_TBBMT = coef(fit3)[1] + coef(fit3)[8]
int3_TCST = coef(fit3)[1] + coef(fit3)[9]


slope3_BT = coef(fit3)[2]
slope3_CKMT = coef(fit3)[2] + coef(fit3)[10]
slope3_CPT = coef(fit3)[2] + coef(fit3)[11]
slope3_EST = coef(fit3)[2] + coef(fit3)[12]
slope3_NSCT = coef(fit3)[2] + coef(fit3)[13]
slope3_NST = coef(fit3)[2] + coef(fit3)[14]
slope3_TBBMT = coef(fit3)[2] + coef(fit3)[15]
slope3_TCST = coef(fit3)[2] + coef(fit3)[16]



```

### Create Dataframe with summry output
```{r}
unique(climate1$EcoCode)
```
```{r}
EcoCode = c("BT", "CKMT", "CPT", "EST", "NSCT", "NST", "TBBMT", "TCST")
intercept = c(int3_BT, int3_CKMT, int3_CPT, int3_EST, int3_NSCT, int3_NST, int3_TBBMT, int3_TCST)
slope = c(slope3_BT, slope3_CKMT, slope3_CPT, slope3_EST, slope3_NSCT, slope3_NST, slope3_TBBMT, slope3_TCST)

sum_Tmmx =data.frame(EcoCode,intercept, slope)
sum_Tmmx
```

### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
climate4 =  climate1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ SummerTmmx, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>%
  mutate(variable="tmmx") %>%
  left_join(sum_Tmmx) %>%
  dplyr::select(variable, EcoCode, intercept, slope, r.squared:nobs)

climate4
```

```{r}
climate1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ SummerTmmx, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(EcoCode), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "Max Temp", y = expression(R^{2}))
```



### Legend Plot
```{r}
legend_bt = ggplot(data = climate1, aes(x = SummerTmmx, y = log10SizeHa, color = BTCodes)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbBTCodesPalette) +
    labs(y=("Annual area burned\n(ha)"), x=(expression(atop(paste("Summer Temperature (", degree,"C)"))))) +
  theme_bw() + 
  theme(legend.position = "right",
        legend.title = element_blank()) +
  guides(colour = guide_legend(ncol = 2))

legend_bt
```

### Combo Plot
```{r fig.height=2.3, fig.width=3}
g3 = ggplot(data = climate1, aes(x = SummerTmmx, y = log10SizeHa, color = BTCodes)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbBTCodesPalette) +
    labs(y=("Annual area burned\n(ha)"), x=(expression(atop(paste("Summer temperature (", degree,"C)"))))) +
  reg_theme

g3
```




## Fire Seaon Model
```{r}
fit4 = lm(log10SizeHa ~ season + EcoCode + season:EcoCode, data = fire1)
```
#### Residuals


```{r}
# Get the fitted value for each observation
fire1$fit4 = fitted(fit4)
```

```{r}
# Get the residuals of the model
fire1$res4 = residuals(fit4)
```

```{r}
qplot(x = fit4, y = res4, data = fire1,
main = "Residuals vs Fitted Values")
```


```{r}
qplot(x = season, y = res4, data = fire1,
xlab = "season",
ylab = "Residuals",
main = "Residuals vs season")
```
```{r}
qplot(x = factor(1), y = res4, data = fire1, geom = "boxplot")
```

```{r}
qqnorm(fire1$res4, main = "Normal QQ Plot of Residuals")
qqline(fire1$res4) # add reference line to the qq plot
```

```{r}
plot(fit4, which = 1) # residual vs fitted values
```

```{r}
plot(fit4, which = 2) # qqnorm plot of residuals
```

### Summary
```{r}
summary(fit4)
```
```{r}
summary(fit4)$r.squared 
```
```{r}
int_BT = coef(fit4)[1]
int_CKMT = coef(fit4)[1] + coef(fit4)[3]
int_CPT = coef(fit4)[1] + coef(fit4)[4]
int_EST = coef(fit4)[1] + coef(fit4)[5]
int_NSCT = coef(fit4)[1] + coef(fit4)[6]
int_NST = coef(fit4)[1] + coef(fit4)[7]
int_TBBMT = coef(fit4)[1] + coef(fit4)[8]
int_TCST = coef(fit4)[1] + coef(fit4)[9]

slope_BT = coef(fit4)[2]
slope_CKMT = coef(fit4)[2] + coef(fit4)[10]
slope_CPT = coef(fit4)[2] + coef(fit4)[11]
slope_EST = coef(fit4)[2] + coef(fit4)[12]
slope_NSCT = coef(fit4)[2] + coef(fit4)[13]
slope_NST = coef(fit4)[2] + coef(fit4)[14]
slope_TBBMT = coef(fit4)[2] + coef(fit4)[15]
slope_TCST = coef(fit4)[2] + coef(fit4)[16]


```

#### Create Dataframe with summry output

```{r}
EcoCode = c("BT", "CKMT", "CPT", "EST", "NSCT", "NST", "TBBMT", "TCST")
intercept = c(int_BT, int_CKMT, int_CPT, int_EST, int_NSCT, int_NST, int_TBBMT, int_TCST)
slope = c(slope_BT, slope_CKMT, slope_CPT, slope_EST, slope_NSCT, slope_NST, slope_TBBMT, slope_TCST)

sum_season =data.frame(EcoCode,intercept, slope)
sum_season
```

#### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
season =  fire1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ season, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results)%>%
  mutate(variable="seasonLength") %>%
  left_join(sum_season) %>%
  dplyr::select(variable, EcoCode, intercept, slope, r.squared:nobs)

season
```



```{r}
fire1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ season, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(EcoCode), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "Ecocode", y = expression(R^{2}))
```


### Combo Plot

```{r fig.height=2.3, fig.width=3}
g4 = ggplot(data = fire1, aes(x = season, y = log10SizeHa, color = BTCodes)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbBTCodesPalette) +
    labs(y=("Annual area burned\n(ha)"), x= "Season length (days)") +
    reg_theme

g4
```



## Snow Model
```{r}
fit5 = lm(log10SizeHa ~ MeanCalDoY + EcoCode + MeanCalDoY:EcoCode, data = snow1)
```

### Residuals


```{r}
# Get the fitted value for each observation
snow$fit5 = fitted(fit5)
```

```{r}
# Get the residuals of the model
snow$res5 = residuals(fit5)
```

```{r}
qplot(x = fit5, y = res5, data = snow,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = MeanCalDoY, y = res5, data = snow,
xlab = "CMD",
ylab = "Residuals",
main = "Residuals vs CMD")
```
```{r}
qplot(x = factor(1), y = res5, data = snow, geom = "boxplot")
```

```{r}
qqnorm(snow$res5, main = "Normal QQ Plot of Residuals")
qqline(snow$res5) # add reference line to the qq plot
```

```{r}
plot(fit5, which = 1) # residual vs fitted values
```

```{r}
plot(fit5, which = 2) # qqnorm plot of residuals
```

### Summary
```{r}
summary(fit5)
```
```{r}
summary(fit5)$r.squared 
```
```{r}
int_BT = coef(fit5)[1]
int_CKMT = coef(fit5)[1] + coef(fit5)[3]
int_CPT = coef(fit5)[1] + coef(fit5)[4]
int_EST = coef(fit5)[1] + coef(fit5)[5]
int_NSCT = coef(fit5)[1] + coef(fit5)[6]
int_NST = coef(fit5)[1] + coef(fit5)[7]
int_TBBMT = coef(fit5)[1] + coef(fit5)[8]
int_TCST = coef(fit5)[1] + coef(fit5)[9]

slope_BT = coef(fit5)[2]
slope_CKMT = coef(fit5)[2] + coef(fit5)[10]
slope_CPT = coef(fit5)[2] + coef(fit5)[11]
slope_EST = coef(fit5)[2] + coef(fit5)[12]
slope_NSCT = coef(fit5)[2] + coef(fit5)[13]
slope_NST = coef(fit5)[2] + coef(fit5)[14]
slope_TBBMT = coef(fit5)[2] + coef(fit5)[15]
slope_TCST = coef(fit5)[2] + coef(fit5)[16]

```

### Create Dataframe with summry output

```{r}
EcoCode = c("BT", "CKMT", "CPT", "EST", "NSCT", "NST", "TBBMT", "TCST")
intercept = c(int_BT, int_CKMT, int_CPT, int_EST, int_NSCT, int_NST, int_TBBMT, int_TCST)
slope = c(slope_BT, slope_CKMT, slope_CPT, slope_EST, slope_NSCT, slope_NST, slope_TBBMT, slope_TCST)

sum_snow =data.frame(EcoCode,intercept, slope)
sum_snow
```

### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
snow2 =  snow1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ MeanCalDoY, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>%
  mutate(variable="SnowOff") %>%
  left_join(sum_snow) %>%
  dplyr::select(variable, EcoCode, intercept, slope, r.squared:nobs)

snow2
```



```{r}
snow1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(log10SizeHa ~ MeanCalDoY, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(EcoCode), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "snow", y = expression(R^{2}))
```


### Combo Plot
```{r fig.height=2.3, fig.width=3}
g5 = ggplot(data = snow1, aes(x = MeanCalDoY, y = log10SizeHa, color = BTCodes)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbBTCodesPalette) +
    labs(y=("Annual area burned\n(ha)"), x= "Snowmelt (Julian date)") +
  reg_theme

g5
```

# Save Summary Tables

```{r}
summer_climate = bind_rows(climate2, climate3, climate4, snow2, season) %>%
  mutate(BTCodes = ifelse(EcoCode =="EST", "B1",
                  ifelse(EcoCode =="NST", "B2",
                   ifelse(EcoCode =="BT", "T1",
                       ifelse(EcoCode =="CPT", "T2",
                         ifelse(EcoCode =="CKMT", "T3",
                            ifelse(EcoCode =="NSCT", "T4",
                              ifelse(EcoCode =="TCST", "T5", "T6")))))))) %>%
  dplyr::select(variable, EcoCode, BTCodes, intercept, slope, adj.r.squared, p.value)

summer_climate
```


```{r eval=FALSE, include=FALSE}
write.csv(summer_climate, '../outputs/ClimateFireAnalysis/TableClimateAnalysis.csv',  row.names = FALSE)
```

# Manuscript Figures 
```{r}
legend1 <- get_legend(legend_bt) 
```

```{r fig.height=7, fig.width=6}
fig1 = cowplot::plot_grid(g1, g4, g2, g5, g3, legend1, ncol=2,  align = "hv") 
   
fig1
```


```{r fig.height=7, fig.width=6}
fig1.1 = fig1  +
  draw_label("(a)", x = 0.03, y = .97, fontfamily = "sans", fontface = "plain", color = "black", size = 12) +
   draw_label("(b)", x = 0.53, y = .97, fontfamily = "sans", fontface = "plain", color = "black", size = 12) +
  draw_label("(c)", x = 0.03, y = .64, fontfamily = "sans", fontface = "plain", color = "black", size = 12) +
  draw_label("(d)", x = 0.53, y = .64, fontfamily = "sans", fontface = "plain", color = "black", size = 12) +
  draw_label("(e)", x = 0.03, y = .33, fontfamily = "sans", fontface = "plain", color = "black", size = 12) 


fig1.1
```

```{r eval=FALSE, include=FALSE}
ggsave("../figures/ClimateAnalysis.png", plot = fig1.1, width = 6, height = 7, units = c("in"), dpi=600 )
```

**THE END**
