==> scripts/3.2-SnowSeasonAnalysis.Rmd <==
---
title: "Season and Snowmelt Analysis"
author: "Anna Talucci"
date: "7/15/2021"
output: html_document
---



# Clear workspace

```{r}
rm(list=ls()) 
```


# Overview
Snow data from MODIS

# Library
```{r include=FALSE}

library(dplyr)
library(tidyr)
library(ggplot2)
library(cowplot)
library(tidyverse)
library(broom)
library(ggplot2)
library(stringr)
library(lme4)
library(cowplot)
library(lme4)
library(lsmeans)
library(magrittr)
library(rcompanion)
library(qpcR)
library(tibble)
```

# Define graph Theme and color Palette 

##Themes
```{r}
lb_theme = theme_bw() + theme(legend.position = "bottom") +  
  theme(panel.grid.major = element_blank()) +
    theme(axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        axis.line = element_line(colour = "black"))
```

```{r}
comp_theme = theme_bw() + theme(legend.position = "none") +  
  theme(panel.grid.major = element_blank()) +
    theme(axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        axis.line = element_line(colour = "black"))
```

### No Legend
```{r}
bar_theme = theme_bw() + 
  theme(legend.position = "none", legend.title = element_blank()) +
  theme(panel.grid.major = element_blank()) +
    theme(axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        axis.line = element_line(colour = "black")) +
  theme(axis.text.x = element_text(angle=65, vjust=0.6)) 
```

```{r}
bar2_theme = theme_bw() + 
  theme(panel.grid.major = element_blank()) +
    theme(axis.title.y = element_text(size = 11, hjust = 0.5, vjust = 1.1),
        axis.text.x = element_text(size = 10, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        axis.line = element_line(colour = "black")) +
  theme(axis.text.x = element_text(angle=65, vjust=0.6)) 
```


## Palettes
```{r}
cbbPalette <- c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
```

```{r}
ecoPalette <- c("#CB453D", "#44CB30", "#2E41C3", "#36A4B4", "#B83491", "#B09C32", "#AE3D5F", "#2DC086", "#3261A1")
```

# Data 

```{r}
data = read.csv("../data/AllFires/SiberiaFires_v2_2001-2020.csv") 
```

```{r}
data
```
## Adjust data
Convert class to factor
```{r}
data = data %>% 
  mutate_at(vars(Class), list(factor)) %>% 
  mutate_at(vars(EcoBiom), list(factor)) %>% 
  filter(EcoBiom!="8")
```

```{r}
unique(data$EcoBiom)
```
```{r}
class(data$Class)
levels(data$Class)
```

```{r}
class(data$EcoBiom)
levels(data$EcoBiom)
```

```{r}
head(data)
```





```{r}
data1 = data %>%
  group_by(FireYr, EcoName, EcoCode) %>%
  summarise(start = round(min(MinDay), 0),
            end = round(max(MaxDay), 0),
            MHa = (sum(SizeHa))/1000000) %>%
  mutate(season = end-start)


data1
   
```

## Analysis
### Model Climate Moisture Deficit
```{r}
fit1 = lm(MHa ~ season + EcoCode + season:EcoCode, data = data1)
```
#### Residuals


```{r}
# Get the fitted value for each observation
data1$fit1 = fitted(fit1)
```

```{r}
# Get the residuals of the model
data1$res1 = residuals(fit1)
```

```{r}
qplot(x = fit1, y = res1, data = data1,
main = "Residuals vs Fitted Values")
```


```{r}
qplot(x = season, y = res1, data = data1,
xlab = "CMD",
ylab = "Residuals",
main = "Residuals vs CMD")
```
```{r}
qplot(x = factor(1), y = res1, data = data1, geom = "boxplot")
```

```{r}
qqnorm(data1$res1, main = "Normal QQ Plot of Residuals")
qqline(data1$res1) # add reference line to the qq plot
```

```{r}
plot(fit1, which = 1) # residual vs fitted values
```

```{r}
plot(fit1, which = 2) # qqnorm plot of residuals
```

### Summary
```{r}
summary(fit1)
```
```{r}
summary(fit1)$r.squared 
```
```{r}
int_BT = coef(fit1)[1]
int_CKMT = coef(fit1)[1] + coef(fit1)[3]
int_CPT = coef(fit1)[1] + coef(fit1)[4]
int_EST = coef(fit1)[1] + coef(fit1)[5]
int_NSCT = coef(fit1)[1] + coef(fit1)[6]
int_NST = coef(fit1)[1] + coef(fit1)[7]
int_TBBMT = coef(fit1)[1] + coef(fit1)[8]
int_TCST = coef(fit1)[1] + coef(fit1)[9]

slope_BT = coef(fit1)[2]
slope_CKMT = coef(fit1)[2] + coef(fit1)[10]
slope_CPT = coef(fit1)[2] + coef(fit1)[11]
slope_EST = coef(fit1)[2] + coef(fit1)[12]
slope_NSCT = coef(fit1)[2] + coef(fit1)[13]
slope_NST = coef(fit1)[2] + coef(fit1)[14]
slope_TBBMT = coef(fit1)[2] + coef(fit1)[15]
slope_TCST = coef(fit1)[2] + coef(fit1)[16]


```

#### Create Dataframe with summry output

```{r}
EcoCode = c("BT", "CKMT", "CPT", "EST", "NSCT", "NST", "TBBMT", "TCST")
intercept = c(int_BT, int_CKMT, int_CPT, int_EST, int_NSCT, int_NST, int_TBBMT, int_TCST)
slope = c(slope_BT, slope_CKMT, slope_CPT, slope_EST, slope_NSCT, slope_NST, slope_TBBMT, slope_TCST)

sum_Tmmx =data.frame(EcoCode,intercept, slope)
sum_Tmmx
```

#### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
season =  data1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(MHa ~ season, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results)%>%
  mutate(variable="seasonLength") %>%
  left_join(sum_Tmmx) %>%
  dplyr::select(variable, EcoCode, intercept, slope, r.squared:nobs)

season
```

```{r}
write.csv(season, '../outputs/ClimateFireAnalysis/TableSeasonLengthAnalysis.csv',  row.names = FALSE)
```

```{r}
data1 %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(MHa ~ season, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(EcoCode), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "Ecocode", y = expression(R^{2}))
```


### Plot

```{r fig.height=2.3, fig.width=6}
plot_season1 = ggplot(data = data1, aes(x = season, y = MHa)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_jitter(size = 1) +
  scale_color_manual(values=cbbPalette) +
    labs(y="Millions of hectares\n(Mha)", x= "Season length (days)") +
  facet_wrap(~EcoCode, scales = "free", ncol=4 ) +
 
scale_x_continuous(guide = guide_axis(check.overlap = TRUE), breaks = scales::pretty_breaks(n=3))+
 scale_y_continuous(guide = guide_axis(check.overlap = TRUE), breaks = scales::pretty_breaks(n=2))+
  theme_bw() +
  theme(strip.background = element_rect(color=NA, fill="white", size=1.5, linetype="solid")) +
  theme(plot.margin = unit(c(.1,.1,.1,.1), "cm"))

plot_season1
  
```



```{r eval=FALSE, include=FALSE}
ggsave("../figures/Analysis/AnalysisSeasonAreaBurned.jpeg", plot = plot_season1, width = 6, height = 2.3, units = c("in"), dpi=600 )
```

```{r}
g4 = ggplot(data = data1, aes(x = season, y = sqrt(MHa), color = EcoCode)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbPalette) +
    labs(y=("Annual Area burned in millions of Ha\n(Squareroot transformation)"), x= "Summer Temperature", color = "EcoCode") +
  theme_bw() + 
  theme(legend.position = "bottom") +
  guides(colour = guide_legend(nrow = 1))

g4
```

# Snow Analysis

## Data 


```{r}
snow = read.csv("../data/SnowData/SnowOff2001-2020.csv") 
```

```{r}
snow
```

# Analysis  Annual Variables
## Model snow Moisture Deficit
```{r}
fit1 = lm(sqrt(Mha) ~ MeanCalDoY + EcoCode + MeanCalDoY:EcoCode, data = snow)
```
#### Residuals


```{r}
# Get the fitted value for each observation
snow$fit1 = fitted(fit1)
```

```{r}
# Get the residuals of the model
snow$res1 = residuals(fit1)
```

```{r}
qplot(x = fit1, y = res1, data = snow,
main = "Residuals vs Fitted Values")
```

```{r}
qplot(x = MeanCalDoY, y = res1, data = snow,
xlab = "CMD",
ylab = "Residuals",
main = "Residuals vs CMD")
```
```{r}
qplot(x = factor(1), y = res1, data = snow, geom = "boxplot")
```

```{r}
qqnorm(snow$res1, main = "Normal QQ Plot of Residuals")
qqline(snow$res1) # add reference line to the qq plot
```

```{r}
plot(fit1, which = 1) # residual vs fitted values
```

```{r}
plot(fit1, which = 2) # qqnorm plot of residuals
```

### Summary
```{r}
summary(fit1)
```
```{r}
summary(fit1)$r.squared 
```
```{r}
int_BT = coef(fit1)[1]
int_CKMT = coef(fit1)[1] + coef(fit1)[3]
int_CPT = coef(fit1)[1] + coef(fit1)[4]
int_EST = coef(fit1)[1] + coef(fit1)[5]
int_NSCT = coef(fit1)[1] + coef(fit1)[6]
int_NST = coef(fit1)[1] + coef(fit1)[7]
int_TBBMT = coef(fit1)[1] + coef(fit1)[8]
int_TCST = coef(fit1)[1] + coef(fit1)[9]

slope_BT = coef(fit1)[2]
slope_CKMT = coef(fit1)[2] + coef(fit1)[10]
slope_CPT = coef(fit1)[2] + coef(fit1)[11]
slope_EST = coef(fit1)[2] + coef(fit1)[12]
slope_NSCT = coef(fit1)[2] + coef(fit1)[13]
slope_NST = coef(fit1)[2] + coef(fit1)[14]
slope_TBBMT = coef(fit1)[2] + coef(fit1)[15]
slope_TCST = coef(fit1)[2] + coef(fit1)[16]

```

# Create Dataframe with summry output

```{r}
EcoCode = c("BT", "CKMT", "CPT", "EST", "NSCT", "NST", "TBBMT", "TCST")
intercept = c(int_BT, int_CKMT, int_CPT, int_EST, int_NSCT, int_NST, int_TBBMT, int_TCST)
slope = c(slope_BT, slope_CKMT, slope_CPT, slope_EST, slope_NSCT, slope_NST, slope_TBBMT, slope_TCST)

sum_Tmmx =data.frame(EcoCode,intercept, slope)
sum_Tmmx
```

### Pulling out r-squared for each group
reference: https://drsimonj.svbtle.com/running-a-model-on-separate-groups

lm(formula = biomass_log ~ ndvi + res_meters + ndvi:res_meters, 
    data = field_ndvi_long1
```{r}
snow1 =  snow %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(Mha ~ MeanCalDoY, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>%
  mutate(variable="SnowOff") %>%
  left_join(sum_Tmmx) %>%
  dplyr::select(variable, EcoCode, intercept, slope, r.squared:nobs)

snow1
```

```{r}
write.csv(snow1, '../outputs/ClimateFireAnalysis/TableSnowOffAnalysis.csv',  row.names = FALSE)
```

```{r}
snow %>% 
  nest(-EcoCode) %>% 
  mutate(fit = map(data, ~ lm(Mha ~ MeanCalDoY, data = .)),
         results = map(fit, glance)) %>% 
  unnest(results) %>% 
  ggplot(aes(x = factor(EcoCode), y = r.squared)) +
    geom_bar(stat = "identity") +
    labs(x = "snow Moisture Deficit", y = expression(R^{2}))
```


### Plot

```{r fig.height=2.3, fig.width=6}
plot_snow1 = ggplot(data = snow, aes(x = MeanCalDoY, y = Mha)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +

  
  labs(y="Millions of hectares\n(Mha)", x= "Snowmelt (Julian date)") +
  facet_wrap(~EcoCode, scales = "free", ncol=4 ) +
 
scale_x_continuous(guide = guide_axis(check.overlap = TRUE), breaks = scales::pretty_breaks(n=3))+
 scale_y_continuous(guide = guide_axis(check.overlap = TRUE), breaks = scales::pretty_breaks(n=2))+
  theme_bw() +
  theme(strip.background = element_rect(color=NA, fill="white", size=1.5, linetype="solid")) +
  theme(plot.margin = unit(c(.1,.1,.1,.1), "cm"))

plot_snow1
  
```

```{r eval=FALSE, include=FALSE}
ggsave("../figures/Analysis/AnalysisSnowAreaBurned.jpeg", plot = plot_snow1, width = 6, height = 2.3, units = c("in"), dpi=600 )
```

```{r}
g5 = ggplot(data = snow, aes(x = MeanCalDoY, y = sqrt(Mha), color = EcoCode)) + 
  geom_smooth(method="lm", se=TRUE, alpha=0.1) +
  geom_point(size = 1) +
  scale_color_manual(values=cbbPalette) +
    labs(y=("Annual Area burned in millions of Ha\n(Squareroot transformation)"), x= "Snowmelt (Julian date)", color = "EcoCode") +
  theme_bw() + 
  theme(legend.position = "bottom") +
  guides(colour = guide_legend(nrow = 1))

g5
```

# Correlation
```{r}
snow %>%
  group_by(EcoCode) %>%
  summarize(COR=cor( MeanCalDoY, Mha))

```

# Combine Plots

#  Combine plots

```{r fig.height=4.7, fig.width=6}
comboplot1 = cowplot::plot_grid(plot_season1, plot_snow1, ncol=1, rel_heights = c(1, 1), align = "hv", labels = c('(a)', '(b)'), label_size = 12, label_fontface = "plain") 
   

comboplot1
```

```{r eval=FALSE, include=FALSE}
ggsave("../figures/Analysis/SeasonSnowTrends.png", plot = comboplot1, width = 6, height = 4.7, units = c("in"), dpi=600 )
```
**THE END**
